SAP-604: Removing redundant home-page.js
